fix(qzp-v2): break coverage_* cyclic-import — extract CoverageStats to coverage_types
Closes 13 CodeQL alerts (12 py/unsafe-cyclic-import + 1 py/cyclic-import)
that traced through the same import structure:

  assert_coverage_100  →  coverage_support
                              ↓
                          coverage_findings + coverage_parsers
                              ↓
                          assert_coverage_100  ← CYCLE

The TYPE_CHECKING guard masked it at runtime, but the analyzer
correctly flagged that whichever module loaded first could see the
other's symbols undefined during the load order race. Three of those
alerts (in coverage_parsers.py functions) were even using runtime
inline imports — same cycle, just lazier.

Fix: extract CoverageStats to a leaf module (coverage_types.py) that
nothing in the coverage stack imports from. Now:

  coverage_types  ← assert_coverage_100, coverage_support,
                     coverage_findings, coverage_parsers
                     (no cycles back into coverage_types)

assert_coverage_100 still re-exports CoverageStats for callers that
already import `from scripts.quality.assert_coverage_100 import
CoverageStats` — backwards compatible.

Also updates sonar-project.properties to add coverage_types.py to
sonar.coverage.exclusions (it's not in tool.coverage.run.source so
the contract test enforces this lockstep).

Tests: 1479 pass (+0). Contract test confirms exclusion list mirror.
Lizard clean (max CCN 3.6). Semgrep clean.

Effect: platform CodeQL alerts drop from 20 → 7 (only the
intentional dual-import test pattern remains, which is design-intent
for monkey-patching).
