+++ sonar-project.properties
@@ -20,4 +20,4 @@ sonar.exclusions=generated/**,docs/admin/**
 # Phase 4 of the QZP rollout (per pyproject.toml comment) is to expand the
 # coverage source to the full scripts/quality/ tree once the rollup_v2 work
 # is stable. When that lands, this exclusion list shrinks accordingly.
-sonar.coverage.exclusions=scripts/__init__.py,scripts/beads-fetch-conversation-history.ts,scripts/beads-fetch-pr-comments.ts,scripts/beads-self-reflect.ts,scripts/quality/__init__.py,scripts/quality/admin_dashboard_pages.py,scripts/quality/alert_dispatch.py,scripts/quality/alert_triggers.py,scripts/quality/alerts.py,scripts/quality/apply_drift_pr.py,scripts/quality/assert_coverage_100.py,scripts/quality/bootstrap_repo.py,scripts/quality/build_admin_dashboard.py,scripts/quality/build_quality_rollup.py,scripts/quality/bump_rollout.py,scripts/quality/bump_workflow_shas.py,scripts/quality/bumps.py,scripts/quality/bypass_labels.py,scripts/quality/check_applitools_zero.py,scripts/quality/check_chromatic_zero.py,scripts/quality/check_codacy_zero.py,scripts/quality/check_codeql_zero.py,scripts/quality/check_deepscan_zero.py,scripts/quality/check_deepsource_zero.py,scripts/quality/check_dependabot_alerts.py,scripts/quality/check_quality_secrets.py,scripts/quality/check_required_checks.py,scripts/quality/check_semgrep_zero.py,scripts/quality/check_sentry_zero.py,scripts/quality/check_sonar_zero.py,scripts/quality/codacy_zero_helpers.py,scripts/quality/codacy_zero_support.py,scripts/quality/common.py,scripts/quality/control_plane.py,scripts/quality/control_plane_admin.py,scripts/quality/control_plane_vendors.py,scripts/quality/coverage_findings.py,scripts/quality/coverage_parsers.py,scripts/quality/coverage_paths.py,scripts/quality/coverage_support.py,scripts/quality/deepsource_html.py,scripts/quality/drift_sync.py,scripts/quality/ensure_codex_auth.py,scripts/quality/export_profile.py,scripts/quality/generate_ruleset_payload.py,scripts/quality/known_issues.py,scripts/quality/marker_regions.py,scripts/quality/normalize_coverage_for_qlty.py,scripts/quality/post_pr_quality_comment.py,scripts/quality/profile_contract_validation.py,scripts/quality/profile_coverage_normalization.py,scripts/quality/profile_normalization.py,scripts/quality/profile_shape.py,scripts/quality/render_codex_prompt.py,scripts/quality/render_repo_baseline.py,scripts/quality/run_codex_exec.py,scripts/quality/run_coverage_gate.py,scripts/quality/run_qlty_zero.py,scripts/quality/run_quality_zero_gate.py,scripts/quality/secrets_sync.py,scripts/quality/security_class_guard.py,scripts/quality/severity_rollup.py,scripts/quality/sonar_zero_support.py,scripts/quality/string_helpers.py,scripts/quality/template_render.py,scripts/quality/validate_codecov_flags.py,scripts/quality/validate_control_plane.py,scripts/quality/verify_v2_deployment.py,scripts/security_helpers.py
+sonar.coverage.exclusions=scripts/__init__.py,scripts/beads-fetch-conversation-history.ts,scripts/beads-fetch-pr-comments.ts,scripts/beads-self-reflect.ts,scripts/quality/__init__.py,scripts/quality/admin_dashboard_pages.py,scripts/quality/alert_dispatch.py,scripts/quality/alert_triggers.py,scripts/quality/alerts.py,scripts/quality/apply_drift_pr.py,scripts/quality/assert_coverage_100.py,scripts/quality/bootstrap_repo.py,scripts/quality/build_admin_dashboard.py,scripts/quality/build_quality_rollup.py,scripts/quality/bump_rollout.py,scripts/quality/bump_workflow_shas.py,scripts/quality/bumps.py,scripts/quality/bypass_labels.py,scripts/quality/check_applitools_zero.py,scripts/quality/check_chromatic_zero.py,scripts/quality/check_codacy_zero.py,scripts/quality/check_codeql_zero.py,scripts/quality/check_deepscan_zero.py,scripts/quality/check_deepsource_zero.py,scripts/quality/check_dependabot_alerts.py,scripts/quality/check_quality_secrets.py,scripts/quality/check_required_checks.py,scripts/quality/check_semgrep_zero.py,scripts/quality/check_sentry_zero.py,scripts/quality/check_sonar_zero.py,scripts/quality/codacy_zero_helpers.py,scripts/quality/codacy_zero_support.py,scripts/quality/common.py,scripts/quality/control_plane.py,scripts/quality/control_plane_admin.py,scripts/quality/control_plane_vendors.py,scripts/quality/coverage_findings.py,scripts/quality/coverage_parsers.py,scripts/quality/coverage_paths.py,scripts/quality/coverage_support.py,scripts/quality/coverage_types.py,scripts/quality/deepsource_html.py,scripts/quality/drift_sync.py,scripts/quality/ensure_codex_auth.py,scripts/quality/export_profile.py,scripts/quality/generate_ruleset_payload.py,scripts/quality/known_issues.py,scripts/quality/marker_regions.py,scripts/quality/normalize_coverage_for_qlty.py,scripts/quality/post_pr_quality_comment.py,scripts/quality/profile_contract_validation.py,scripts/quality/profile_coverage_normalization.py,scripts/quality/profile_normalization.py,scripts/quality/profile_shape.py,scripts/quality/render_codex_prompt.py,scripts/quality/render_repo_baseline.py,scripts/quality/run_codex_exec.py,scripts/quality/run_coverage_gate.py,scripts/quality/run_qlty_zero.py,scripts/quality/run_quality_zero_gate.py,scripts/quality/secrets_sync.py,scripts/quality/security_class_guard.py,scripts/quality/severity_rollup.py,scripts/quality/sonar_zero_support.py,scripts/quality/string_helpers.py,scripts/quality/template_render.py,scripts/quality/validate_codecov_flags.py,scripts/quality/validate_control_plane.py,scripts/quality/verify_v2_deployment.py,scripts/security_helpers.py
